Refactor: items fatch 코드 api폴더로 이동
