+++ src/app/items/[itemId]/detail/page.tsx
@@ -1,9 +1,7 @@
 import Image from 'next/image'
 import React from 'react'
 
-import { fetchItemDetails } from '../layout'
-
-// http://3.36.91.126:8080/api/items?itemId=1
+import { fetchItemDetails } from '@/api/resource/items'
 
 export default async function page({ params }: { params: { itemId: number } }) {
   const itemDetails = await fetchItemDetails(params.itemId)
