+++ src/app/items/[itemId]/layout.tsx
@@ -1,16 +1,9 @@
 import React from 'react'
 
+import { fetchItemDetails } from '@/api/resource/items'
 import BottomTab from '@/components/feature/item/BottomTab'
 import ItemHeader from '@/components/feature/item/ItemHeader'
 import ItemTabs from '@/components/feature/item/ItemTabs'
-
-export async function fetchItemDetails(itemId: number) {
-  const res = await fetch(`http://3.36.91.126:8080/api/items?itemId=${itemId}`)
-  if (!res.ok) {
-    throw new Error('Failed to fetch data')
-  }
-  return await res.json()
-}
 
 export default async function Itemlayout({
   children,
